Refactor and enhance code for better readability and performance
- Updated variable names for clarity in WebLazyAssets.cpp and WebPlatform.cpp.
- Changed data types for the Dead attribute in w_CharacterInfo.h from BYTE to float for improved precision.
- Added new frame timing functions in ZzzScene.h to better track performance metrics.
- Improved error logging and path handling in WebLazyAssets.cpp.
- Refactored canvas size handling in WebTitleScene.cpp for consistency and clarity.
- Adjusted function signatures in multiple files to use more descriptive parameter names.

+++ source/w_CharacterInfo.h
@@ -175,7 +175,7 @@ public:
     BYTE        AttackFlag;
 
     BYTE        TargetAngle;
-    BYTE        Dead;
+    float       Dead;
     WORD		Skill;
     BYTE        SwordCount;
     BYTE		byExtensionSkill;
